feat(tui): o `:sync` mostra a volta do laço e as duas pontas onde o áudio some
Uma linha nova, e é a que responde "é a rede ou é esta máquina?":

    LAÇO volta 22.4ms · reposição 1 · reacerto 0 · recusa 0 · anel 0

`volta` é o pior tempo entre duas conferidas do prazo de reprodução. Abaixo de
20 ms a máquina acompanha o relógio sozinha; acima, só está em dia porque a
reposição a segura. `recusa` e `anel` são o que não saiu para o transporte e o
que não chegou ao dispositivo — as duas eram `let _ =`.

+++ crates/seele-audio/src/playout.rs
@@ -133,8 +133,8 @@ impl PlayoutClock {
         }
 
         // `period` nunca é zero — ver `new`.
-        let elapsed = u128::from(late.as_nanos());
-        let period = u128::from(self.period.as_nanos()).max(1);
+        let elapsed = late.as_nanos();
+        let period = self.period.as_nanos().max(1);
         let overdue = u32::try_from(elapsed / period).unwrap_or(u32::MAX);
         let frames = overdue.saturating_add(1);
